Influence Function for the Quintile Share Ratio (QSR)

Computes the linearized variable (influence function) for the quintile share ratio (QSR) using the linearization approach based on Deville (1999) and the derivation in Langel and Tillé (2011) .

if_qsr(y, weights = NULL, type = 4)

Arguments

y

A numeric vector of income or other continuous variable.

weights

A numeric vector of sampling weights. If NULL (default), equal weights are assumed for all observations.

type

Quantile estimation type: integer 4–9 or "HD" for Harrell–Davis (default: 4)

Value

A numeric vector of the same length as y containing the linearized variable \(\widehat{z}_k\) for each observation.

Details

According to the definition by Langel and Tillé (2011) , the influence function of QSR is given by:

$$I(\mathrm{QSR})_k=\frac{y_k-I\left(Y_{0.8}\right)_k}{Y_{0.2}}-\frac{\left(Y-Y_{0.8}\right) I\left(Y_{0.2}\right)_k}{Y_{0.2}^2}$$

where \(I(Y_p)_k = p Q(p) - (Q(p) - y_k) \mathbb{1}[y_k \leq Q(p)]\) is the influence function of the partial total up to quantile \(p\) and \(Y_{\alpha} = \sum_{j} y_j\mathbb{1}[y_k \leq {Q}(\alpha)]\).

The estimated linearized variable is: $$\widehat{z}_k=\frac{y_k-\left\{0.8 \widehat{Q}(0.8)-\left(\widehat{Q}(0.8)-y_k\right) \mathbb{1}\left[y_k \leq \widehat{Q}(0.8)\right]\right\}}{\widehat{Y}_{0.2}}-\frac{\left(\widehat{Y}-\widehat{Y}_{0.8}\right)\left\{0.2 \widehat{Q}(0.2)-\left(\widehat{Q}(0.2)-y_k\right) \mathbb{1}\left[y_k \leq \widehat{Q}(0.2)\right]\right\}}{\widehat{Y}_{0.2}^2}$$

where \(\hat{Y}_{\alpha} = \sum_{j \in s} w_j y_j\mathbb{1}[y_k \leq \widehat{Q}(\alpha)]\) and \(\widehat{Q}(p)\) is the weighted p-th quantile estimator, computed using the internal function csquantile().

References

Deville J (1999). “Variance estimation for complex statistics and estimators: linearization and residual techniques.” Survey methodology, 25, 193–204.

Langel M, Tillé Y (2011). “Statistical inference for the quintile share ratio.” Journal of Statistical Planning and Inference, 141, 2976–2985.

See also

Examples

data(synthouse)

eq <- synthouse$eq_income ### Income data

# Simple example
z <- if_qsr(eq)

# With weights
w <- synthouse$weight
z_weighted <- if_qsr(y = eq, weights = w)